+++ frontend/src/styles/components/_post.scss
@@ -127,6 +127,7 @@
   &__action-icon {
     width: 34px;
     height: 34px;
+    cursor: pointer;
     path {
       stroke-width: 2.2px;
